Add LamC and ClosV definitions
This is the beginning of function support. A LamC represents a lambda or
anonymous function with a list of parameter names and an ExprC body.
When a LamC is interpreted, the resulting Value is a ClosV, which
contains the same param list and body, but also the current Env of the
interpreter.

By taking taking the env at the time of function definition instead of
function application lets the functions have lexical scope, which is
prety cool.