Add function application
Function application is represented as an AppC, which contains an ExprC
for the function expression and a list of ExprCs for arguments.

To interpret a function application, the function ExprC is interpreted
into a ClosV, and each argument is interpreted and added to the new
environment that the ClosV's ExprC body is interpreted with.